delete scope.py file
